# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tech Dev API",
        "description": "API desenvolvida para o desafio da Gazin-tech",
        "contact": {
            "email": "caiogrechic50@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
        "/api/desenvolvedores": {
            "get": {
                "tags": [
                    "Desenvolvedores"
                ],
                "summary": "Lista todos os desenvolvedores",
                "description": "Retorna uma lista paginada de desenvolvedores",
                "operationId": "1ba3c5aa3f1a25a3ec17a1244bd40f15",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número da página",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Número de itens por página",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Campo que vai ser ordenado",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Ordenação crescente (asc) ou descrescente (desc)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Filtra pelo id do desenvolvedor",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "nome",
                        "in": "query",
                        "description": "Filtra pelo nome do desenvolvedor",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "hobby",
                        "in": "query",
                        "description": "Filtra pelo hobby do desenvolvedor",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sexo",
                        "in": "query",
                        "description": "Filtra pelo sexo do desenvolvedor",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "nivel_id",
                        "in": "query",
                        "description": "Filtra pelo nivel do desenvolvedor",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de desenvolvedores",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "description": "ID do desenvolvedor",
                                                        "type": "integer"
                                                    },
                                                    "nome": {
                                                        "description": "Nome do desenvolvedor",
                                                        "type": "string"
                                                    },
                                                    "hobby": {
                                                        "description": "Hobby do desenvolvedor",
                                                        "type": "string"
                                                    },
                                                    "sexo": {
                                                        "description": "Sexo do desenvolvedor",
                                                        "type": "string"
                                                    },
                                                    "data_nascimento": {
                                                        "description": "Data de nascimento do desenvolvedor",
                                                        "type": "string",
                                                        "format": "date"
                                                    },
                                                    "nivel_id": {
                                                        "description": "ID do nível do desenvolvedor",
                                                        "type": "integer"
                                                    },
                                                    "idade": {
                                                        "description": "Idade do desenvolvedor",
                                                        "type": "integer",
                                                        "readOnly": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "total": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Requisição inválida"
                    }
                }
            },
            "post": {
                "tags": [
                    "Desenvolvedores"
                ],
                "summary": "Cria um novo desenvolvedor",
                "description": "Cria um novo desenvolvedor",
                "operationId": "5d450d92e34ec4faf658a7186ecb995c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "schema": "Desenvolvedor",
                                "title": "Desenvolvedor",
                                "properties": {
                                    "search": {
                                        "description": "Nome do desenvolvedor",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Desenvolvedor criado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "schema": "Desenvolvedor",
                                    "title": "Desenvolvedor",
                                    "required": [
                                        "nome",
                                        "hobby",
                                        "nivel_id",
                                        "sexo",
                                        "data_nascimento"
                                    ],
                                    "properties": {
                                        "id": {
                                            "description": "ID do desenvolvedor",
                                            "type": "integer"
                                        },
                                        "nome": {
                                            "description": "Nome do desenvolvedor",
                                            "type": "string"
                                        },
                                        "hobby": {
                                            "description": "Hobby do desenvolvedor",
                                            "type": "string"
                                        },
                                        "sexo": {
                                            "description": "Sexo do desenvolvedor",
                                            "type": "string"
                                        },
                                        "data_nascimento": {
                                            "description": "Data de nascimento do desenvolvedor",
                                            "type": "string",
                                            "format": "date"
                                        },
                                        "nivel_id": {
                                            "description": "ID do nível do desenvolvedor",
                                            "type": "integer"
                                        },
                                        "idade": {
                                            "description": "Idade do desenvolvedor",
                                            "type": "integer",
                                            "readOnly": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Requisição inválida"
                    }
                }
            }
        },
...
    },
...
}